+++ Sandswept/Survivors/Ranger/SkillDefs/Passive/OverchargedProtection.cs
@@ -5,7 +5,7 @@ namespace Sandswept.Survivors.Ranger.SkillDefs.Passive
     public class OverchargedProtection : SkillBase<OverchargedProtection>
     {
         public override string Name => "Overcharged Protection";
-        public override string Description => "Hold up to " + Projectiles.DirectCurrent.maxCharge + " $rcCharge$ec. $rcCharge$ec increases $shhealth regeneration$se by up to $sh2.5 hp/s$se and $sharmor$se by up to $sh15$se. $rcCharge decays over time$ec.".AutoFormat();
+        public override string Description => "Hold up to " + Projectiles.DirectCurrent.maxCharge + " $rcCharge$ec. $rcCharge$ec increases $shhealth regeneration$se by up to $sh2.5 hp/s$se and $sharmor$se by up to $sh15$se. $suCharge decays over time$se.".AutoFormat();
         public override Type ActivationStateType => typeof(GenericCharacterMain);
         public override string ActivationMachineName => "Body";
         public override float Cooldown => 0f;
@@ -70,11 +70,12 @@ namespace Sandswept.Survivors.Ranger.SkillDefs.Passive
                 {
                     regenAccumulator += chargeRegen * Time.fixedDeltaTime;
 
-                    if (regenAccumulator >= 1f) {
+                    if (regenAccumulator >= 1f)
+                    {
                         ProcChainMask mask = new();
                         mask.AddModdedProc(OverchargeRegen);
                         hc.Heal(regenAccumulator, mask, false);
-                        
+
                         regenAccumulator = 0f;
                     }
                 }