+++ Sandswept/Survivors/Ranger/SkillDefs/Secondary/Char.cs
@@ -6,7 +6,7 @@ namespace Sandswept.Survivors.Ranger.SkillDefs.Secondary
     {
         public override string Name => "Char";
 
-        public override string Description => "$sdIgnite$se. $suReduce current heat by 50%$se. Fire off a blazing orb for $sd1200%$se damage that $sdengulfs$se the ground on impact for $sd250%$se damage per second. $suDamage increases with heat spent$se.".AutoFormat();
+        public override string Description => "$sdIgnite$se. $suReduce current heat by 50%$se. Fire off a blazing orb for $sd200%$se damage that $sdengulfs$se the ground on impact for $sd200% damage per second$se. $suDamage increases with heat spent$se.".AutoFormat();
 
         public override Type ActivationStateType => typeof(States.Secondary.Char);